refactor: 테스트 코드 구조 개선 및 도메인별 분리
- 서비스 및 레포지토리 테스트 코드를 도메인별로 분리하여 응집력 강화
  - 기존에는 service 및 repository 폴더에 모든 테스트 코드가 혼재
  - 도메인별로 테스트 코드를 정리하여 관련성 높은 코드끼리 묶음
