+++ src/test/java/com/example/solidconnection/unit/board/repository/BoardRepositoryTest.java
@@ -1,10 +1,8 @@
-package com.example.solidconnection.unit.repository;
+package com.example.solidconnection.unit.board.repository;
 
 import com.example.solidconnection.board.domain.Board;
 import com.example.solidconnection.board.repository.BoardRepository;
 import com.example.solidconnection.custom.exception.CustomException;
-import com.example.solidconnection.post.domain.PostLike;
-import com.example.solidconnection.post.repository.PostLikeRepository;
 import com.example.solidconnection.post.domain.Post;
 import com.example.solidconnection.post.repository.PostRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
@@ -13,6 +11,7 @@ import com.example.solidconnection.type.Gender;
 import com.example.solidconnection.type.PostCategory;
 import com.example.solidconnection.type.PreparationStatus;
 import com.example.solidconnection.type.Role;
+import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,15 +20,14 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.example.solidconnection.custom.exception.ErrorCode.INVALID_POST_LIKE;
+import static com.example.solidconnection.custom.exception.ErrorCode.INVALID_BOARD_CODE;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 @ActiveProfiles("test")
-@DisplayName("게시글 좋아요 레포지토리 테스트")
-class PostLikeRepositoryTest {
+@DisplayName("게시판 레포지토리 테스트")
+class BoardRepositoryTest {
     @Autowired
     private PostRepository postRepository;
     @Autowired
@@ -37,24 +35,30 @@ class PostLikeRepositoryTest {
     @Autowired
     private SiteUserRepository siteUserRepository;
     @Autowired
-    private PostLikeRepository postLikeRepository;
+    private EntityManager entityManager;
 
-    private Post post;
     private Board board;
     private SiteUser siteUser;
-    private PostLike postLike;
-
+    private Post post;
 
     @BeforeEach
-    void setUp() {
+    public void setUp() {
         board = createBoard();
         boardRepository.save(board);
+
         siteUser = createSiteUser();
         siteUserRepository.save(siteUser);
+
         post = createPost(board, siteUser);
         post = postRepository.save(post);
-        postLike = createPostLike(post, siteUser);
-        postLikeRepository.save(postLike);
+
+        entityManager.flush();
+        entityManager.clear();
+    }
+
+    private Board createBoard() {
+        return new Board(
+                "FREE", "자유게시판");
     }
 
     private SiteUser createSiteUser() {
@@ -67,11 +71,6 @@ class PostLikeRepositoryTest {
                 Role.MENTEE,
                 Gender.MALE
         );
-    }
-
-    private Board createBoard() {
-        return new Board(
-                "FREE", "자유게시판");
     }
 
     private Post createPost(Board board, SiteUser siteUser) {
@@ -87,36 +86,56 @@ class PostLikeRepositoryTest {
         return post;
     }
 
-    private PostLike createPostLike(Post post, SiteUser siteUser) {
-        PostLike postLike = new PostLike();
-        postLike.setPostAndSiteUser(post, siteUser);
-        return postLike;
-    }
-
     @Test
     @Transactional
-    void 게시글_좋아요를_조회한다() {
+    public void 게시판을_조회할_때_게시글은_즉시_로딩한다() {
         // when
-        PostLike foundPostLike = postLikeRepository.getByPostAndSiteUser(post, siteUser);
+        Board foundBoard = boardRepository.getByCodeUsingEntityGraph(board.getCode());
+        foundBoard.getPostList().size(); // 추가쿼리 발생하지 않는다.
 
         // then
-        assertEquals(foundPostLike, postLike);
+        assertThat(foundBoard.getCode()).isEqualTo(board.getCode());
     }
 
     @Test
     @Transactional
-    void 게시글_좋아요를_조회할_때_유효한_좋아요가_아니라면_예외_응답을_반환한다() {
+    public void 게시판을_조회할_때_게시글은_즉시_로딩한다_유효한_게시판이_아니라면_예외_응답을_반환한다() {
         // given
-        postLike.resetPostAndSiteUser();
-        postLikeRepository.delete(postLike);
+        String invalidCode = "INVALID_CODE";
 
         // when, then
         CustomException exception = assertThrows(CustomException.class, () -> {
-            postLikeRepository.getByPostAndSiteUser(post, siteUser);
+            boardRepository.getByCodeUsingEntityGraph(invalidCode);
         });
         assertThat(exception.getMessage())
-                .isEqualTo(INVALID_POST_LIKE.getMessage());
+                .isEqualTo(INVALID_BOARD_CODE.getMessage());
         assertThat(exception.getCode())
-                .isEqualTo(INVALID_POST_LIKE.getCode());
+                .isEqualTo(INVALID_BOARD_CODE.getCode());
+    }
+
+    @Test
+    @Transactional
+    public void 게시판을_조회한다() {
+        // when
+        Board foundBoard = boardRepository.getByCode(board.getCode());
+
+        // then
+        assertEquals(board.getCode(), foundBoard.getCode());
+    }
+
+    @Test
+    @Transactional
+    public void 게시판을_조회할_때_유효한_게시판이_아니라면_예외_응답을_반환한다() {
+        // given
+        String invalidCode = "INVALID_CODE";
+
+        // when, then
+        CustomException exception = assertThrows(CustomException.class, () -> {
+            boardRepository.getByCode(invalidCode);
+        });
+        assertThat(exception.getMessage())
+                .isEqualTo(INVALID_BOARD_CODE.getMessage());
+        assertThat(exception.getCode())
+                .isEqualTo(INVALID_BOARD_CODE.getCode());
     }
 }
