+++ src/test/java/com/example/solidconnection/unit/comment/service/CommentServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.unit.service;
+package com.example.solidconnection.unit.comment.service;
 
 import com.example.solidconnection.board.domain.Board;
 import com.example.solidconnection.comment.domain.Comment;
