+++ src/test/java/com/example/solidconnection/unit/post/repository/PostLikeRepositoryTest.java
@@ -1,10 +1,10 @@
-package com.example.solidconnection.unit.repository;
+package com.example.solidconnection.unit.post.repository;
 
 import com.example.solidconnection.board.domain.Board;
 import com.example.solidconnection.board.repository.BoardRepository;
-import com.example.solidconnection.comment.domain.Comment;
-import com.example.solidconnection.comment.repository.CommentRepository;
 import com.example.solidconnection.custom.exception.CustomException;
+import com.example.solidconnection.post.domain.PostLike;
+import com.example.solidconnection.post.repository.PostLikeRepository;
 import com.example.solidconnection.post.domain.Post;
 import com.example.solidconnection.post.repository.PostRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
@@ -17,22 +17,19 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
-import static com.example.solidconnection.custom.exception.ErrorCode.INVALID_COMMENT_ID;
+import static com.example.solidconnection.custom.exception.ErrorCode.INVALID_POST_LIKE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-
-@SpringBootTest
-@ActiveProfiles("dev")
-@DisplayName("댓글 레포지토리 테스트")
-class CommentRepositoryTest {
+@DataJpaTest
+@ActiveProfiles("test")
+@DisplayName("게시글 좋아요 레포지토리 테스트")
+class PostLikeRepositoryTest {
     @Autowired
     private PostRepository postRepository;
     @Autowired
@@ -40,34 +37,24 @@ class CommentRepositoryTest {
     @Autowired
     private SiteUserRepository siteUserRepository;
     @Autowired
-    private CommentRepository commentRepository;
+    private PostLikeRepository postLikeRepository;
 
+    private Post post;
     private Board board;
     private SiteUser siteUser;
-    private Post post;
-    private Comment parentComment;
-    private Comment childComment;
+    private PostLike postLike;
+
 
     @BeforeEach
-    public void setUp() {
+    void setUp() {
         board = createBoard();
         boardRepository.save(board);
-
         siteUser = createSiteUser();
         siteUserRepository.save(siteUser);
-
         post = createPost(board, siteUser);
         post = postRepository.save(post);
-
-        parentComment = createParentComment();
-        childComment = createChildComment();
-        commentRepository.save(parentComment);
-        commentRepository.save(childComment);
-    }
-
-    private Board createBoard() {
-        return new Board(
-                "FREE", "자유게시판");
+        postLike = createPostLike(post, siteUser);
+        postLikeRepository.save(postLike);
     }
 
     private SiteUser createSiteUser() {
@@ -80,6 +67,11 @@ class CommentRepositoryTest {
                 Role.MENTEE,
                 Gender.MALE
         );
+    }
+
+    private Board createBoard() {
+        return new Board(
+                "FREE", "자유게시판");
     }
 
     private Post createPost(Board board, SiteUser siteUser) {
@@ -95,56 +87,36 @@ class CommentRepositoryTest {
         return post;
     }
 
-    private Comment createParentComment() {
-        Comment comment = new Comment(
-                "parent"
-        );
-        comment.setPostAndSiteUser(post, siteUser);
-        return comment;
-    }
-
-    private Comment createChildComment() {
-        Comment comment = new Comment(
-                "child"
-        );
-        comment.setParentCommentAndPostAndSiteUser(parentComment, post, siteUser);
-        return comment;
+    private PostLike createPostLike(Post post, SiteUser siteUser) {
+        PostLike postLike = new PostLike();
+        postLike.setPostAndSiteUser(post, siteUser);
+        return postLike;
     }
 
     @Test
     @Transactional
-    public void 재귀쿼리로_댓글트리를_조회한다() {
+    void 게시글_좋아요를_조회한다() {
         // when
-        List<Comment> commentTreeByPostId = commentRepository.findCommentTreeByPostId(post.getId());
+        PostLike foundPostLike = postLikeRepository.getByPostAndSiteUser(post, siteUser);
 
         // then
-        List<Comment> expectedResponse = List.of(parentComment, childComment);
-        assertEquals(commentTreeByPostId, expectedResponse);
+        assertEquals(foundPostLike, postLike);
     }
 
     @Test
     @Transactional
-    public void 댓글을_조회한다() {
-        // when
-        Comment foundComment = commentRepository.getById(parentComment.getId());
-
-        // then
-        assertEquals(parentComment, foundComment);
-    }
-
-    @Test
-    @Transactional
-    public void 댓글을_조회할_때_유효한_댓글이_아니라면_예외_응답을_반환한다() {
+    void 게시글_좋아요를_조회할_때_유효한_좋아요가_아니라면_예외_응답을_반환한다() {
         // given
-        Long invalidId = -1L;
+        postLike.resetPostAndSiteUser();
+        postLikeRepository.delete(postLike);
 
         // when, then
         CustomException exception = assertThrows(CustomException.class, () -> {
-            commentRepository.getById(invalidId);
+            postLikeRepository.getByPostAndSiteUser(post, siteUser);
         });
         assertThat(exception.getMessage())
-                .isEqualTo(INVALID_COMMENT_ID.getMessage());
+                .isEqualTo(INVALID_POST_LIKE.getMessage());
         assertThat(exception.getCode())
-                .isEqualTo(INVALID_COMMENT_ID.getCode());
+                .isEqualTo(INVALID_POST_LIKE.getCode());
     }
 }
